smart_brief_generator/requirements.txt
google-generativeai
python-dotenv
pydantic
fpdf2
